Cleanup Rules from __init__
